fix: fix type in MainActivity and refactore the file auth to authentication
