+++ app/src/main/java/com/github/se/signify/MainActivity.kt
@@ -53,7 +53,7 @@ class MainActivity : ComponentActivity() {
 
   private lateinit var sharedPreferencesTheme: SharedPreferences
   private lateinit var sharedPreferencesLanguage: SharedPreferences
-  // Now when launching the MainActivity, the builder will now which dependency provider to use
+  // Now when launching the MainActivity, the builder will know which dependency provider to use
   // (Test or Base)
   private val dependencyProvider by lazy { (application as BaseApplication).dependencyProvider }
 
@@ -161,7 +161,8 @@ fun SignifyAppPreview(
               } else {
                 navigationActions.navigateTo(Screen.TUTORIAL)
               }
-            })
+            },
+            dependencyProvider.provideAuthService())
       }
       composable(Screen.UNAUTHENTICATED.route) { UnauthenticatedScreen(navigationActions) }
     }
